HGImporter.process_ref: always reset to head.rev()
Formerly, the git ref would be reset to head.rev() only if the range was
non-empty, even if all revs were already marked, but would reset to the
latest unmarked rev otherwise (which should match for well-behaved hg
repos).  I suspect that this handling of empty ranges was a cause of
problems similar to issue #13 that I have experienced, and that more
reliable logic is to always reset git's ref to head.rev().