set HGRCPATH='' in both tests and remote.
This disallows loading of system or the user's personal hg config files, such as ~/.hgrc.
The main motivator for this is that extensions the user has loaded can cause dissatisfaction
for both scripts and tests. The script will complain bitterly, but will not communicate
clearly if it encounters an extension doing unexpected things -- unexpected by the script,
that is. We can assume the extension thinks it knows what it is doing.

All tests pass with this change, but I am marginally concerned about side effects in which
things like not having the user's credentials set could cause explosions.