Add bin script for testing.
Testing gitifyhg is difficult when there's an installed egg version: eggs
are placed on sys.path ahead of anything in PYTHONPATH.  This commit adds
a bin script that puts a source-code gitifyhg on sys.path ahead of
everything else, so it can be tested.  It is NOT intended that this replace
the bin script installed by setup.py; it is only for testing.  With this
script, I can say:
PATH=/path/to/gitifyhg/bin git ...
and get the version of gitifyhg in /path/to/gitifyhg/gitifyhg.
To ensure this bin script isn't used unintentionally, it's named
git-remote-testgitifyhg, so you have to also have a remote called:
testgitifyhg::<mercurial/repo>
i.e., you can have separate production and development gitifyhg remotes.