Log version, and output when requested.
Introducing the use of the ArgumentParser to handle the argument parsing.
This should not get in the way of any communication from git, because the
manpage for git-remote-helpers says:

SYNOPSIS
       git remote-<transport> <repository> [<URL>]

i.e., there are no -- or - options provided, so as long as all additional
arguments (e.g., --version) are prepended with hyphens, there's no worry,
and we can just dump all the other arguments into the HGRemote.