Fix this monstrosity of a test (test_special_cases.t)
Not only did it not actually duplicate the py.test test it referred to,
it also failed to set LC_ALL (which meant the filesystemencoding was
ANSI_X3.4-1968).  As well, I also fixed some testing-specific idioms
(like having to specify the user when you clone a repo or when you make
an hg commit).