test: don't empty HGRCPATH
Extensions can't interfere, because we need to enable them in our hgrc,
which we specify manually so no other hgrc interferes.

This would allow to test that HGRCPATH is working properly inside
gitifyhg.py, if it decides to read the configuration using the
heuristics of Mercurial (see git-remote-hg for reference), or that it's
properly isolated from a real environment that, which would need a
separate test that pollutes ~/.hgrc.

Comments by Jed Brown.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>