Allow HGRCPATH from environment to propagate to mercurial
If the user sets HGRCPATH, it is likely because they want to affect the
environment used to make commits (relevant when translating git
lightweight tags to mercurial tags) or because they would like to add an
extension such as keyring or schemes to the environment.

We are not aware of Mercurial extensions that would negatively affect
the behavior of the interfaces that Gitifyhg uses.

This reverts part of 468a4fe8c46f7fb6f571e1902b91b1ce66b831ea

Suggested-by: Felipe Contreras <felipe.contreras@gmail.com>